reserved shell definitions
