+++ mods/server/server.go
@@ -8,7 +8,6 @@ import (
 	"crypto/rsa"
 	"crypto/x509"
 	"encoding/hex"
-	"encoding/json"
 	"encoding/pem"
 	"fmt"
 	"io"
@@ -503,7 +502,7 @@ func (s *svr) Start() error {
 			sshd.OptionAuthServer(s),
 			sshd.OptionGrpcServerAddress(s.conf.Grpc.Listeners...),
 			sshd.OptionMotdMessage(fmt.Sprintf("machbase-neo %s %s", mods.VersionString(), mods.Edition())),
-			sshd.OptionShellDefinitionProvider(s.GetShellDef),
+			sshd.OptionCustomShellProvider(s.GetSshShell),
 		)
 		if err != nil {
 			return errors.Wrap(err, "shell server")
@@ -803,87 +802,6 @@ func (s *svr) checkListenPort(address string) error {
 	return nil
 }
 
-func (s *svr) IterateShellDefs(cb func(*sshd.ShellDefinition) bool) error {
-	if cb == nil {
-		return nil
-	}
-	entries, err := os.ReadDir(s.shellDefsDir)
-	if err != nil {
-		return nil
-	}
-	for _, entry := range entries {
-		if !strings.HasSuffix(entry.Name(), ".json") || entry.IsDir() {
-			continue
-		}
-		content, err := os.ReadFile(filepath.Join(s.shellDefsDir, entry.Name()))
-		if err != nil {
-			s.log.Errorf("ERR file access, %s", err.Error())
-			continue
-		}
-		def := &sshd.ShellDefinition{}
-		if err := json.Unmarshal(content, def); err != nil {
-			s.log.Warnf("ERR invalid shell conf, %s", err.Error())
-			continue
-		}
-		def.Name = strings.ToUpper(strings.TrimSuffix(entry.Name(), ".json"))
-		shouldContinue := cb(def)
-		if !shouldContinue {
-			break
-		}
-	}
-	return nil
-}
-
-func (s *svr) SetShellDef(def *sshd.ShellDefinition) error {
-	name := strings.ToUpper(def.Name)
-	if name == "SHELL" {
-		return fmt.Errorf("'%s' is not allowed for the custom shell name", name)
-	}
-	if len(def.Args) == 0 {
-		return errors.New("invalid command for the custom shell")
-	}
-	binpath := def.Args[0]
-	if fi, err := os.Stat(binpath); err != nil {
-		return errors.Wrapf(err, "'%s' is not accessible", binpath)
-	} else {
-		if fi.IsDir() {
-			return fmt.Errorf("'%s' is not executable", binpath)
-		}
-		if runtime.GOOS == "windows" {
-			if !strings.HasSuffix(strings.ToLower(binpath), ".exe") && !strings.HasSuffix(strings.ToLower(binpath), ".com") {
-				return fmt.Errorf("'%s' is not executable", binpath)
-			}
-		} else {
-			if fi.Mode().Perm()&0111 == 0 {
-				return fmt.Errorf("'%s' is not executable", binpath)
-			}
-		}
-	}
-	content, err := json.Marshal(def)
-	if err != nil {
-		return err
-	}
-	path := filepath.Join(s.shellDefsDir, fmt.Sprintf("%s.json", name))
-	return os.WriteFile(path, content, 0600)
-}
-
-func (s *svr) RemoveShellDef(name string) error {
-	path := filepath.Join(s.shellDefsDir, fmt.Sprintf("%s.json", strings.ToUpper(name)))
-	return os.Remove(path)
-}
-
-func (s *svr) GetShellDef(name string) (found *sshd.ShellDefinition) {
-	name = strings.ToUpper(name)
-	s.IterateShellDefs(func(def *sshd.ShellDefinition) bool {
-		if def.Name == name {
-			found = def
-			return false
-		}
-		return true
-	})
-	return
-}
-
 // AuthorizedCertificate returns client's X.509 certificate, it returns nil if not found with the given id
 func (s *svr) AuthorizedCertificate(id string) (*x509.Certificate, error) {
 	path := filepath.Join(s.authorizedKeysDir, fmt.Sprintf("%s_cert.pem", id))
@@ -1095,71 +1013,4 @@ func (s *svr) ValidateSshPublicKey(keyType string, key string) (bool, error) {
 		}
 	}
 	return false, nil
-}
-
-// web options
-
-func (s *svr) WebReferences() []httpd.WebReferenceGroup {
-	ret := []httpd.WebReferenceGroup{}
-
-	references := httpd.WebReferenceGroup{Label: "References"}
-	references.Items = append(references.Items, httpd.ReferenceItem{Type: "url", Title: "machbase-neo docs", Addr: "https://neo.machbase.com/", Target: "_blank"})
-	references.Items = append(references.Items, httpd.ReferenceItem{Type: "url", Title: "machbase sql reference", Addr: "http://endoc.machbase.com/", Target: "_blank"})
-	references.Items = append(references.Items, httpd.ReferenceItem{Type: "url", Title: "https://machbase.com", Addr: "https://machbase.com/", Target: "_blank"})
-	ret = append(ret, references)
-
-	tutorials := httpd.WebReferenceGroup{Label: "Tutorials"}
-	tutorials.Items = append(tutorials.Items, httpd.ReferenceItem{Type: "wrk", Title: "Waves in TQL", Addr: "./tutorials/waves_in_tql.wrk"})
-	tutorials.Items = append(tutorials.Items, httpd.ReferenceItem{Type: "wrk", Title: "Fast Fourier Transform in TQL", Addr: "./tutorials/fft_in_tql.wrk"})
-	ret = append(ret, tutorials)
-
-	samples := httpd.WebReferenceGroup{Label: "Samples"}
-	samples.Items = append(samples.Items, httpd.ReferenceItem{Type: "wrk", Title: "markdown cheatsheet", Addr: "./tutorials/sample_markdown.wrk"})
-	samples.Items = append(samples.Items, httpd.ReferenceItem{Type: "wrk", Title: "mermaid cheatsheet", Addr: "./tutorials/sample_mermaid.wrk"})
-	samples.Items = append(samples.Items, httpd.ReferenceItem{Type: "wrk", Title: "pikchr cheatsheet", Addr: "./tutorials/sample_pikchr.wrk"})
-	samples.Items = append(samples.Items, httpd.ReferenceItem{Type: "tql", Title: "user script in tql (1)", Addr: "./tutorials/user-script1.tql"})
-	samples.Items = append(samples.Items, httpd.ReferenceItem{Type: "tql", Title: "user script in tql (2)", Addr: "./tutorials/user-script2.tql"})
-	ret = append(ret, samples)
-
-	return ret
-}
-
-func (s *svr) GetAllWebShells() []*httpd.WebShell {
-	var ret []*httpd.WebShell
-	ret = append(ret, &httpd.WebShell{Type: "sql", Label: "SQL", Icon: "file-document-outline", Id: "SQL"})
-	ret = append(ret, &httpd.WebShell{Type: "tql", Label: "TQL", Icon: "chart-scatter-plot", Id: "TQL"})
-	ret = append(ret, &httpd.WebShell{Type: "wrk", Label: "WORKSHEET", Icon: "clipboard-text-play-outline", Id: "WRK"})
-	ret = append(ret, &httpd.WebShell{Type: "taz", Label: "TAG ANALYZER", Icon: "chart-line", Id: "TAZ"})
-	ret = append(ret, &httpd.WebShell{Type: "term", Label: "SHELL", Icon: "console", Id: "SHELL",
-		Attributes: []httpd.WebShellAttribute{&httpd.WebShellCloneable{Cloneable: true}},
-	})
-	s.IterateShellDefs(func(def *sshd.ShellDefinition) bool {
-		ret = append(ret, &httpd.WebShell{
-			Type:    "term",
-			Label:   def.Name,
-			Icon:    "console-network-outline",
-			Id:      strings.ToUpper(def.Name),
-			Content: strings.Join(def.Args, " "),
-			Attributes: []httpd.WebShellAttribute{
-				&httpd.WebShellCloneable{Cloneable: true},
-				&httpd.WebShellRemovable{Removable: true},
-				&httpd.WebShellEditable{Editable: true},
-			},
-		})
-		return true
-	})
-	return ret
-}
-
-func (s *svr) GetWebShell(id string) (*httpd.WebShell, error) {
-	return nil, nil
-}
-func (s *svr) CopyWebShell(id string) (*httpd.WebShell, error) {
-	return nil, nil
-}
-func (s *svr) RemoveWebShell(id string) (*httpd.WebShell, error) {
-	return nil, nil
-}
-func (s *svr) UpdateWebShell(sh *httpd.WebShell) error {
-	return nil
 }
