--- mods/service/sshd/shell.go
@@ -1,6 +1,0 @@
-package sshd
-
-type ShellDefinition struct {
-	Name string   `json:"-"`
-	Args []string `json:"args"`
-}
