+++ mods/service/sshd/ssh_shell.go
@@ -2,64 +2,47 @@ package sshd
 
 import (
 	"fmt"
-	"os"
-	"runtime"
 	"strings"
 
 	"github.com/gliderlabs/ssh"
 )
+
+type Shell struct {
+	Cmd  string
+	Args []string
+	Envs map[string]string
+}
+
+type ShellDefinition struct {
+	Name string   `json:"-"`
+	Args []string `json:"args"`
+
+	Attributes map[string]string `json:"attributes,omitempty"`
+}
 
 func (svr *sshd) motdProvider(user string) string {
 	return fmt.Sprintf("Greetings, %s\r\n%s\r\n",
 		strings.ToUpper(user), svr.motdMessage)
 }
 
-func (svr *sshd) findShellDefinition(ss ssh.Session) (string, *ShellDefinition) {
+func (svr *sshd) findShell(ss ssh.Session) (string, *Shell) {
 	user := ss.User()
-	var shellDef *ShellDefinition
+	var shell *Shell
 	if strings.Contains(user, ":") {
 		userShell := ""
 		toks := strings.SplitN(user, ":", 2)
 		user = toks[0]
 		userShell = toks[1]
-		shellDef = svr.shellDefinitionProvider(userShell)
-	}
-	return user, shellDef
-}
-
-func (svr *sshd) buildShell(user string, shellDef *ShellDefinition) *Shell {
-	var shell *Shell
-	if shellDef == nil {
-		shell = svr.shellProvider(user)
+		if userShell == "SHELL" {
+			shell = svr.shellProvider(user)
+		} else {
+			shell = svr.customShellProvider(userShell)
+		}
+		if shell == nil {
+			return user, nil
+		}
 	} else {
-		shell = &Shell{}
-		shell.Cmd = shellDef.Args[0]
-		if len(shellDef.Args) > 1 {
-			shell.Args = shellDef.Args[1:]
-		}
+		shell = svr.shellProvider(user)
 	}
-
-	shell.Envs = map[string]string{}
-	if runtime.GOOS == "windows" {
-		envs := os.Environ()
-		for _, line := range envs {
-			if !strings.Contains(line, "=") {
-				continue
-			}
-			toks := strings.SplitN(line, "=", 2)
-			if len(toks) != 2 {
-				continue
-			}
-			shell.Envs[strings.TrimSpace(toks[0])] = strings.TrimSpace(toks[1])
-		}
-		if _, ok := shell.Envs["USERPROFILE"]; !ok {
-			userHomeDir, err := os.UserHomeDir()
-			if err != nil {
-				userHomeDir = "."
-			}
-			shell.Envs["USERPROFILE"] = userHomeDir
-		}
-	}
-
-	return shell
+	return user, shell
 }
